+++ pypdf/_encryption.py
@@ -759,7 +759,7 @@ class PasswordType(IntEnum):
     OWNER_PASSWORD = 2
 
 
-class EncryptAlgorithm(tuple, Enum):  # noqa: SLOT001
+class EncryptAlgorithm(tuple, Enum):  # type: ignore # noqa: SLOT001
     # V, R, Length
     RC4_40 = (1, 2, 40)
     RC4_128 = (2, 3, 128)
@@ -1144,7 +1144,7 @@ class Encryption:
     def make(
         alg: EncryptAlgorithm, permissions: int, first_id_entry: bytes
     ) -> "Encryption":
-        alg_ver, alg_rev, key_bits = cast(tuple, alg)
+        alg_ver, alg_rev, key_bits = alg
 
         stm_filter, str_filter, ef_filter = "/V2", "/V2", "/V2"