+++ tests/test_writer.py
@@ -4,10 +4,12 @@ import shutil
 import subprocess
 from io import BytesIO
 from pathlib import Path
+from typing import Any
 
 import pytest
 
 from pypdf import (
+    ImageType,
     ObjectDeletionFlag,
     PageObject,
     PdfMerger,
@@ -1859,6 +1861,116 @@ def test_object_contains_indirect_reference_to_self():
     outpage = writer.add_blank_page(width, height)
     outpage.merge_page(reader.pages[6])
     writer.append(reader)
+
+
+def test_remove_image_per_type():
+    writer = PdfWriter(clone_from=RESOURCE_ROOT / "reportlab-inline-image.pdf")
+    writer.remove_images(ImageType.INLINE_IMAGES)
+
+    assert all(
+        x not in writer.pages[0].get_contents().get_data()
+        for x in (b"BI", b"ID", b"EI")
+    )
+
+    with pytest.raises(DeprecationWarning):
+        writer.remove_images(True)
+
+    writer = PdfWriter(clone_from=RESOURCE_ROOT / "GeoBase_NHNC1_Data_Model_UML_EN.pdf")
+    writer.remove_images(ImageType.DRAWING_IMAGES)
+    assert all(
+        x not in writer.pages[1].get_contents().get_data()
+        for x in (b" re\n", b"W*", b"f*")
+    )
+    assert all(
+        x in writer.pages[1].get_contents().get_data() for x in (b" TJ\n", b"rg", b"Tm")
+    )
+    assert all(
+        x not in writer.pages[9]["/Resources"]["/XObject"]["/Meta84"].get_data()
+        for x in (b" re\n", b"W*", b"f*")
+    )
+    writer.remove_images(ImageType.XOBJECT_IMAGES)
+    assert b"Do\n" not in writer.pages[0].get_contents().get_data()
+    assert len(writer.pages[0]["/Resources"]["/XObject"]) == 0
+
+
+@pytest.mark.enable_socket()
+def test_add_outlines_on_empty_dict():
+    """Cf #2233"""
+
+    def _get_parent_bookmark(current_indent, history_indent, bookmarks) -> Any:
+        """The parent of A is the nearest bookmark whose indent is smaller than A's"""
+        assert len(history_indent) == len(bookmarks)
+        if current_indent == 0:
+            return None
+        for i in range(len(history_indent) - 1, -1, -1):
+            # len(history_indent) - 1   ===>   0
+            if history_indent[i] < current_indent:
+                return bookmarks[i]
+        return None
+
+    bookmark_lines = """1 FUNDAMENTALS OF RADIATIVE TRANSFER 1
+1.1 The Electromagnetic Spectrum; Elementary Properties of Radiation 1
+1.2 Radiative Flux 2
+    Macroscopic Description of the Propagation of Radiation 2
+    Flux from an Isotropic Source-The Inverse Square Law 2
+1.3 The Specific Intensity and Its Moments 3
+    Definition of Specific Intensity or Brightness 3
+    Net Flux and Momentum Flux 4
+    Radiative Energy Density 5
+    Radiation Pressure in an Enclosure Containing an Isotropic Radiation Field 6
+    Constancy of Specific Zntensiw Along Rays in Free Space 7
+    Proof of the Inverse Square Law for a Uniformly Bright Sphere 7
+1.4 Radiative Transfer 8
+    Emission 9
+    Absorption 9
+    The Radiative Transfer Equation 11
+    Optical Depth and Source Function 12
+    Mean Free Path 14
+    Radiation Force 15
+1.5 Thermal Radiation 15
+    Blackbody Radiation 15
+    Kirchhof's Law for Thermal Emission 16
+    Thermodynamics of Blackbody Radiation 17
+    The Planck Spectrum 20
+    Properties of the Planck Law 23
+    Characteristic Temperatures Related to Planck Spectrum 25
+1.6 The Einstein Coefficients 27
+    Definition of Coefficients 27
+    Relations between Einstein Coefficients 29
+    Absorption and Emission Coefficients in Terms of Einstein Coefficients 30
+1.7 Scattering Effects; Random Walks 33
+    Pure Scattering 33
+    Combined Scattering and Absorption 36
+1.8 Radiative Diffusion 39
+    The Rosseland Approximation 39
+    The Eddington Approximation; Two-Stream Approximation 42
+PROBLEMS 45
+REFERENCES 50
+2 BASIC THEORY OF RADIATION FIELDS 51
+2.1 Review of Maxwell’s Equations 51
+2.2 Plane Electromagnetic Waves 55
+2.3 The Radiation Spectrum 58
+2.4 Polarization and Stokes Parameters 62
+    Monochromatic Waves 62
+    Quasi-monochromatic Waves 65
+2.5 Electromagnetic Potentials 69
+2.6 Applicability of Transfer Theory and the Geometrical Optics Limit 72
+PROBLEMS 74
+REFERENCES 76"""
+    url = "https://github.com/py-pdf/pypdf/files/12797067/test-12.pdf"
+    name = "iss2233.pdf"
+    reader = PdfReader(BytesIO(get_data_from_url(url, name=name)))
+    writer = PdfWriter(clone_from=reader)
+
+    bookmarks, history_indent = [], []
+    for line in bookmark_lines.split("\n"):
+        line2 = re.split(r"\s+", line.strip())
+        indent_size = len(line) - len(line.lstrip())
+        parent = _get_parent_bookmark(indent_size, history_indent, bookmarks)
+        history_indent.append(indent_size)
+        title, page = " ".join(line2[:-1]), int(line2[-1]) - 1
+        new_bookmark = writer.add_outline_item(title, page, parent=parent)
+        bookmarks.append(new_bookmark)
 
 
 def test_merging_many_temporary_files():